feat: generate logic mapper
